pdf作成を実装

+++ lib/app/domain/entities/daily_work_for_pdf.dart
@@ -7,6 +7,6 @@ class DailyWorkForPDF with _$DailyWorkForPDF {
     required String productName,
     required DateTime workDate,
     required String workDetail,
-    required double workTimeByWorkDetail,
+    required int workTimeByWorkDetail,
   }) = _DailyWorkForPDF;
 }
